feat: workspace name form 분리
